# Make DispatcherQueueSynchronizationContext faster and zero-alloc