--- a/src/Projections/Reunion/Microsoft.UI.Dispatching.DispatcherQueueSynchronizationContext.cs
+++ b/src/Projections/Reunion/Microsoft.UI.Dispatching.DispatcherQueueSynchronizationContext.cs
@@ -1,5 +1,9 @@
 using System;
 using System.Threading;
+using WinRT;
+
+#nullable enable
+#pragma warning disable CA1416
 
 namespace Microsoft.System
 {
@@ -7,7 +11,7 @@ namespace Microsoft.System
     /// DispatcherQueueSyncContext allows developers to await calls and get back onto the
     /// UI thread. Needs to be installed on the UI thread through DispatcherQueueSyncContext.SetForCurrentThread
     /// </summary>
-    public class DispatcherQueueSynchronizationContext : SynchronizationContext
+    public partial class DispatcherQueueSynchronizationContext : SynchronizationContext
     {
         private readonly DispatcherQueue m_dispatcherQueue;
 
@@ -17,7 +21,7 @@ namespace Microsoft.System
         }
 
         /// <inheritdoc/>
-        public override void Post(SendOrPostCallback d, object state)
+        public override unsafe void Post(SendOrPostCallback d, object? state)
         {
 #if NET6_0
             ArgumentNullException.ThrowIfNull(d, nameof(d));
@@ -33,11 +37,35 @@ namespace Microsoft.System
             }
 #endif
 
-            m_dispatcherQueue.TryEnqueue(() => d(state));
+#if NET5_0_OR_GREATER
+            DispatcherQueueProxyHandler* dispatcherQueueProxyHandler = DispatcherQueueProxyHandler.Create(d!, state);
+            bool success;
+            int hResult;
+
+            try
+            {
+                IDispatcherQueue* dispatcherQueue = (IDispatcherQueue*)((IWinRTObject)m_dispatcherQueue).NativeObject.ThisPtr;
+
+                hResult = dispatcherQueue->TryEnqueue(dispatcherQueueProxyHandler, (byte*)&success);
+
+                GC.KeepAlive(this);
+            }
+            finally
+            {
+                dispatcherQueueProxyHandler->Release();
+            }
+
+            if (hResult != 0)
+            {
+                ExceptionHelpers.ThrowExceptionForHR(hResult);
+            }
+#else
+            m_dispatcherQueue.TryEnqueue(() => d!(state));
+#endif
         }
 
         /// <inheritdoc/>
-        public override void Send(SendOrPostCallback d, object state)
+        public override void Send(SendOrPostCallback d, object? state)
         {
             throw new NotSupportedException("Send not supported");
         }