… src/Projections/Reunion/Reunion.csproj