## Overview

The current `DispatcherQueueSynchronizationContext` implementation is rather inefficient, for two reasons:
- It's creating a delegate with closure every single time (!), capturing both the callback and the state
- There's also additional overhead in general due to how the actual callback is then scheduled internally

Here's the specific line that's troubling:

https://github.com/microsoft/CsWinRT/blob/31a191d0e9b557e3732f93d86cf626defbb8d009/src/Projections/Reunion/Microsoft.UI.Dispatching.DispatcherQueueSynchronizationContext.cs#L25

This type is used extremely frequently in applications, especially for those that eg. rely on portable libraries for things such as MVVM support etc. that also deal with proper dispatching. The way that works is by capturing the synchronization context during setup and then dispatch through that, meaning this `Post` method will get invoked every time.

## The proposed solution

This PR optimizes the `Post` implementation by making it faster and zero-alloc.
There are no public API changes.
Opening this as draft as I'm not entirely sure of the procedure to add tests for this for this specific repo.
Pinging @jkoritzinsky as we discussed this feature idea previously on Discord as well 😄

I run some benchmarks for this a while back, got some nice numbers:

|                 Method |     Mean |     Error |    StdDev | Ratio |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------------- |---------:|----------:|----------:|------:|-------:|------:|------:|----------:|
|  TryEnqueueWithClosure | 7.062 us | 0.1353 us | 0.1852 us |  1.00 | 0.0687 |     - |     - |     312 B |
|    **TryEnqueueWithState** | **1.139 us** | 0.0229 us | 0.0676 us |  **0.15** |      - |     - |     - |         **-** |

In this scenario dispatching with this system was about 7x faster, and also completely allocation free.
I'd argue it'd be nice to have such a basic building block like this one to be as efficient as possible.

I'm sure it might be a good idea to setup up some new ones in this repo as well, though I'd need guidance on that 😅

Opening as draft for now because of this, can switch this once everyone's happy with coverage and overall setup.

## Related issues/PRs

- https://github.com/CommunityToolkit/WindowsCommunityToolkit/pull/4097
- https://github.com/microsoft/microsoft-ui-xaml/issues/3321