@manodasanW 2021-12-01
Thanks @Sergio0694 for this contribution.  A ~7x improvement with reduced memory pressure is definitely great to see in something that can be frequently used.  The version of this type that you have edited here is actually a copy of the one in WinAppSDK and is here for test purposes.  Adding @jeffstall and @ionvasilian who own this type and should be able to consider and update the actual one in WinAppSDK.

I looked over the change and it looks good to me.  It basically looks like you use a natively allocated proxy instance to hold onto the callback and state and make it look like a delegate by implementing Invoke and then fetch them later when invoked to make the call.  The only thing I wonder is if the delegate proxy should also implement IReferenceTrackerTarget.  I am not sure if XAML reference tracking is used for this type or not, but Jeff or Ion might be able to comment on that.