@Sergio0694 2021-12-01
Hi @manodasanW, thank you for taking a look, really glad to hear this is something that could be interesting to have! 🎉
I wasn't aware this type was just a copy and not the real one (whoops!), but then again the main point of this PR was to show a proof of concept, share some numbers and start a conversation about this. Happy to help migrate this if needed! 😄

> "I looked over the change and it looks good to me. It basically looks like you use a natively allocated proxy instance to hold onto the callback and state and make it look like a delegate by implementing Invoke and then fetch them later when invoked to make the call."

Yup, that's correct! The underlying WinRT API for `IDispatcherQueue::TryEnqueue` is just this one:

```cpp
int TryEnqueue(IDispatcherQueueHandler* callback, byte* result);
```

Where that `callback` is this an object inheriting from `IDispatcherQueueHandler`, which is:

```cpp
class IDispatcherQueueHandler : IUnknown
{
    int Invoke();
}
```

So in this PR I'm just creating my own implementation instead of letting CsWinRT create one wrapping that managed delegate/closure (we're essentially defining a C# struct that implements an object derived from `IDispatcherQueueHandler`). This means I can then also make it blittable and just store a couple GC handles in it to track the managed objects to invoke later (delegate and optional state), without extra allocations. And the `Invoke()` method which is then invoked by the WinRT runtime when the callback is dispatched simply grabs the target callback and state back and invokes them, and then ensures exceptions are tracked correctly, if any. That `IReferenceTrackerTarget` interface hasn't come up while investigating this so far (as in, `IDispatcherQueueHandler` didn't inherit from it, and when testing this and checking with a breakpoint on `QueryInterface`, that interface was never queried), but for sure, waiting for Jeff to confirm this or suggest any changes 😊