@benwestprog 2022-10-20
We ultimately decided not to port these changes into WinAppSDK. They add considerable code complexity, and we could not justify the added risk.

---
In reply to: [1285755982](https://github.com/microsoft/CsWinRT/pull/1058#issuecomment-1285755982) [](http://example.com/codeflow?ancestors=1285755982,1285755982)